Add path drawing API
This allows the creation of arbitrary shapes via a sequence of lines and
curves. This implementation aims to support almost all of the path
drawing functionality that PDF 1.3 supports, though there are some
minor limitations imposed by the API structure in an effort to make it
more difficult to produce degenerate paths (for example, the API only
supports one clipping path per painted path).

Technically, this is a functional superset of the other drawing
primitives provided by FPDF and they could be implemented using this
backend. However, to minimize the number of changes when introducing a
new feature, none of the existing APIs have been reworked.

This ignores some of the existing APIs that fpdf provides for e.g.
setting the stroke/fill color (though it will default to the colors
set by those API calls) in order to provide more descriptive objects
to accomplish the same purpose.