drawing: rewrite to support alpha blending and style inheriting
To start with, the rendering pass has been updated to pass more
information from one stage to the next. This will allow better support
for some SVG drawing operators that we might as well try to support
because it doesn't hurt anything. In addition, a lot of rework has gone
into how the styling is structured (theoretically with little change to
the API) to allow for automatic determination of painting style from
the style attributes.

Some of the style items are being offloaded into the graphics state
parameter dictionaries which notably allows us to support the PDF
transparency model---this was completely missing from the initial
implementation. Unfortunately, the use of these dicts is a mess, and I
wish they didn't exist. Drawing transparency is a PDF 1.4+ feature, so
we will need to bump the version flag above 1.3 (the default) if using
any of the drawing features, to be sure we don't emit an out-of-spec
PDF.

There are still a number of features we don't support (patterns etc.)
that are not going to be added right now, though the state dict support
should theoretically make it easier to add support in the future,
maybe.

reimplement remaining style attributes

this includes blend modes.

Also, I've just determined that the graphics state parameter dictionary
ordering is nondeterministic thanks to my very skillful use of sets,
which do not preserve order. I will need to swap that implementation
out for an OrderedDict, I think.

commit message quality going downhill

Reintroduce the render tree output, renamed as debug output. This is
now generated during the rendering process, so it has access to the
contextual information used during the rendering process, which allows
even more useful output to be produced, such as the resolved absolute
output of relative path components. I tried doing this with separate
render paths for debug and not debug to avoid adding a bunch of debug
branching, but this ended up kind of falling apart at the
GraphicsContext level, as there's sufficiently entangled logic in the
rendering process that I didn't feel comfortable copy pasting it in.
At the end of the day, I haven't measured the performance of this at
all anyway, so it's kind of silly to try to worry about such
optimizations. However, the real motivation was to avoid polluting the
normal code with a bunch of debug branches, and I'm sad to say I've
failed in this regard. I may come back to this, maybe not.

The graphics state dictionary output should be deterministic now, as
it's been switched over to using OrderedDicts instead of sets.

Adding docstrings is an ongoing process, but the next step will
probably be rewrite the existing unit tests and start adding new ones.

The features are now in place to produce accurate output of the svg
logo test, including properly omitting the stroke and including
transparency.

also add arc support

Now including prettier debug information as well.